requirements.txt
# Release Automation PoC Dependencies

# Jira API
requests>=2.28.0

# Google APIs
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.80.0

# Slack API
slack-sdk>=3.19.0

# Flask (for Slack webhook server)
flask>=2.3.0

# Anthropic Claude API (for LLM consolidation)
anthropic>=0.18.0

# Environment variables
python-dotenv>=1.0.0

# Date/time handling
python-dateutil>=2.8.2

# Scheduling (for local development)
schedule>=1.1.0

# Advanced Scheduling (APScheduler for production)
apscheduler>=3.10.0
pytz>=2023.3
